+++ src/main/java/module/playeranalysis/skillCompare/PlayerComparePanel.java
@@ -13,6 +13,7 @@ import core.model.UserParameter;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.*;
+import java.io.Serial;
 import java.util.List;
 import java.util.Vector;
 
@@ -21,6 +22,7 @@ import java.util.Vector;
  */
 public class PlayerComparePanel extends LazyImagePanel implements ItemListener, FocusListener {
 
+	@Serial
 	private static final long serialVersionUID = -1629490436656226196L;
 	// Members for Tables
 	private PlayerTable m_jTableTop; // Table for all players
@@ -35,8 +37,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 	private JScrollPane m_scrollPanePlayer;
 	private JScrollPane m_scrollPanePlayerGesamt;
 	private PlayerTableModel m_playerTableModelTop;
-	private PlayerTableModel m_playerTableModelBottom;
-	private PlayerTableModel m_playerTableModelDetail;
+    private PlayerTableModel m_playerTableModelDetail;
 	private static JComboBox m_CB_type;
 	private JComboBox m_CB_Experience;
 	private JComboBox m_CB_Form;
@@ -62,17 +63,14 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 	private JComboBox m_CB_Nr_SetPieces;
 	private JComboBox m_CB_Nr_Loyalty;
 	private Vector<module.playeranalysis.skillCompare.Player> m_V_setPlayers;
-	private List<core.model.player.Player> m_V_allPlayers;
-	private module.playeranalysis.skillCompare.Player[] m_ar_allPlayers;
-	private module.playeranalysis.skillCompare.Player[] m_ar_setPlayers;
-	private CBItem[] m_rating = PlayerAbility.ITEMS;
+    private module.playeranalysis.skillCompare.Player[] m_ar_allPlayers;
+    private final CBItem[] m_rating = PlayerAbility.ITEMS;
 	private static int m_selectedRow;
 	private int m_i_ptmTopCount;
-	private int m_numberOfPlayers;
-	private static int[] newRating;
+    private static int[] newRating;
 	private static int[] changedRating;
 	private boolean m_b_refresh = true;
-	private Color lightblue = new Color(235, 235, 255);
+	private final Color lightblue = new Color(235, 235, 255);
 
 	@Override
 	protected void initialize() {
@@ -105,21 +103,9 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 	}
 
 	private void addListeners() {
-		m_btCompare.addActionListener(new ActionListener() {
+		m_btCompare.addActionListener(e -> comparePlayer());
 
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				comparePlayer();
-			}
-		});
-
-		m_btReset.addActionListener(new ActionListener() {
-
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				reset();
-			}
-		});
+		m_btReset.addActionListener(e -> reset());
 	}
 
 	private void initComponents() {
@@ -281,8 +267,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		gbc.gridwidth = GridBagConstraints.REMAINDER;
 		gbl.setConstraints(m_CB_type, gbc);
 		// *******************************
-		gbc.gridwidth = 4;
-		gbc.gridwidth = GridBagConstraints.REMAINDER;
+        gbc.gridwidth = GridBagConstraints.REMAINDER;
 		gbl.setConstraints(m_L_Header, gbc);
 		gbc.gridwidth = 1;
 		// *******************************
@@ -355,17 +340,14 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		// *******
 		gbc.gridwidth = 1;
 		gbl.setConstraints(m_L_HomeGrown, gbc);
-		gbc.gridwidth = 1;
-		gbc.gridwidth = GridBagConstraints.REMAINDER; // end row
+        gbc.gridwidth = GridBagConstraints.REMAINDER; // end row
 		gbl.setConstraints(m_CB_Homegrown, gbc);
 		// *******
-		gbc.gridwidth = 1;
-		// *******************************
+        // *******************************
 		gbc.gridwidth = GridBagConstraints.REMAINDER; // end row
 		gbc.insets = new Insets(10, 3, 0, 3);
 		gbl.setConstraints(m_btCompare, gbc);
-		gbc.gridwidth = 1;
-		// *******************************
+        // *******************************
 		gbc.gridwidth = GridBagConstraints.REMAINDER; // end row
 		gbc.insets = new Insets(0, 3, 0, 3);
 		gbl.setConstraints(m_btReset, gbc);
@@ -463,12 +445,9 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		if (ie.getSource().equals(m_CB_type)) {
 			int cbType = m_CB_type.getSelectedIndex();
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				int spielerID = (Integer) m_playerTableModelTop.getValueAt(i,
-                        m_playerTableModelTop.getColumnCount() - 1);
-				int pos = HOVerwaltung.instance().getModel().getCurrentPlayer(spielerID)
-						.getIdealPosition();
-				String group = HOVerwaltung.instance().getModel().getCurrentPlayer(spielerID)
-						.getTeamGroup();
+				int spielerID = (Integer) m_playerTableModelTop.getValueAt(i,m_playerTableModelTop.getColumnCount() - 1);
+				int pos = HOVerwaltung.instance().getModel().getCurrentPlayer(spielerID).getIdealPosition();
+				String group = HOVerwaltung.instance().getModel().getCurrentPlayer(spielerID).getTeamGroup();
 				// System.out.println(cbType +":"+group);
 				if (cbType == 1 && pos == 0 || cbType == 2 && (pos > 0 && pos < 8) || cbType == 3
 						&& (pos > 7 && pos < 12) || cbType == 4 && (pos > 11 && pos < 16)
@@ -643,7 +622,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		m_scrollPaneTableBottom.setViewportView(null);
 
 		for (int i = 0; i < m_i_ptmTopCount; i++) {
-			if ((Boolean) m_playerTableModelTop.getValueAt(i, 0) == true) {
+			if ((Boolean) m_playerTableModelTop.getValueAt(i, 0)) {
 				m_playerTableModelTop.setValueAt(Boolean.FALSE, i, 0);
 			}
 		}
@@ -657,8 +636,8 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 	 * 
 	 */
 	private void getAllPlayers() {
-		m_V_allPlayers = HOVerwaltung.instance().getModel().getCurrentPlayers();
-		m_numberOfPlayers = m_V_allPlayers.size();
+        List<core.model.player.Player> m_V_allPlayers = HOVerwaltung.instance().getModel().getCurrentPlayers();
+        int m_numberOfPlayers = m_V_allPlayers.size();
 		m_ar_allPlayers = new module.playeranalysis.skillCompare.Player[m_numberOfPlayers];
 		int i=0;
 		for ( core.model.player.Player p : m_V_allPlayers){
@@ -691,7 +670,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		switch (selectedType) {
 		case 0:
 			for (int i = 0; i < m_i_ptmTopCount; i++)
-				if ((Boolean) m_playerTableModelTop.getValueAt(i, 0) == true)
+				if ((Boolean) m_playerTableModelTop.getValueAt(i, 0))
 					fetchPlayer((Integer) m_playerTableModelTop.getValueAt(i,
                             m_playerTableModelTop.getColumnCount() - 1));
 			break;
@@ -701,7 +680,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 				try {
 					tmpPos = ((Float) m_playerTableModelTop.getValueAt(i, 4)).byteValue();
-				} catch (Exception ex) {
+				} catch (Exception ignored) {
 				}
 				if (tmpPos == 0 && m_playerTableModelTop.getValueAt(i, 0) == Boolean.TRUE) {
 					fetchPlayer((Integer) m_playerTableModelTop.getValueAt(i,
@@ -717,7 +696,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 				try {
 					tmpPos = ((Float) m_playerTableModelTop.getValueAt(i, 4)).byteValue();
-				} catch (Exception ex) {
+				} catch (Exception ignored) {
 				}
 
 				if ((tmpPos > 0 && tmpPos < 8)
@@ -736,7 +715,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 				try {
 					tmpPos = ((Float) m_playerTableModelTop.getValueAt(i, 4)).byteValue();
-				} catch (Exception ex) {
+				} catch (Exception ignored) {
 				}
 				if ((tmpPos > 7 && tmpPos < 12)
 						&& m_playerTableModelTop.getValueAt(i, 0) == Boolean.TRUE) {
@@ -755,7 +734,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 				try {
 					tmpPos = ((Float) m_playerTableModelTop.getValueAt(i, 4)).byteValue();
-				} catch (Exception ex) {
+				} catch (Exception ignored) {
 				}
 				if ((tmpPos > 11 && tmpPos < 16)
 						&& m_playerTableModelTop.getValueAt(i, 0) == Boolean.TRUE) {
@@ -775,7 +754,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 				try {
 					tmpPos = ((Float) m_playerTableModelTop.getValueAt(i, 4)).byteValue();
-				} catch (Exception ex) {
+				} catch (Exception ignored) {
 				}
 				if ((tmpPos > 15 && tmpPos < 18)
 						&& m_playerTableModelTop.getValueAt(i, 0) == Boolean.TRUE) {
@@ -791,7 +770,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		case 6:
 
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				String gruppe = "";
+				String gruppe;
 
 				gruppe = m_playerTableModelTop.getValueAt(i, 5).toString();
 				if (gruppe.equals("A-Team")
@@ -808,7 +787,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		case 7:
 
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				String gruppe = "";
+				String gruppe;
 
 				gruppe = m_playerTableModelTop.getValueAt(i, 5).toString();
 				if (gruppe.equals("B-Team")
@@ -825,7 +804,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		case 8:
 
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				String gruppe = "";
+				String gruppe;
 
 				gruppe = m_playerTableModelTop.getValueAt(i, 5).toString();
 				if (gruppe.equals("C-Team")
@@ -842,7 +821,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 		case 9:
 
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				String gruppe = "";
+				String gruppe;
 
 				gruppe = m_playerTableModelTop.getValueAt(i, 5).toString();
 				if (gruppe.equals("D-Team")
@@ -858,7 +837,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 
 		case 10:
 			for (int i = 0; i < m_i_ptmTopCount; i++) {
-				String gruppe = "";
+				String gruppe;
 
 				gruppe = m_playerTableModelTop.getValueAt(i, 5).toString();
 				if (gruppe.equals("E-Team")
@@ -879,13 +858,13 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 			}
 		}
 
-		// Create array from a tablemodel vector
-		m_ar_setPlayers = new module.playeranalysis.skillCompare.Player[m_V_setPlayers.size()];
+		// Create array from a table model vector
+        Player[] m_ar_setPlayers = new Player[m_V_setPlayers.size()];
 		for (int counter = 0; counter < m_ar_setPlayers.length; counter++) {
 			m_ar_setPlayers[counter] = m_V_setPlayers.elementAt(counter);
 		}
 
-		m_playerTableModelBottom = new PlayerTableModel(m_ar_setPlayers, 2);
+        PlayerTableModel m_playerTableModelBottom = new PlayerTableModel(m_ar_setPlayers, 2);
 		TableSorter sorter2 = new TableSorter(m_playerTableModelBottom);
 		m_jTableBottom = new PlayerTable(sorter2, m_playerTableModelBottom);
 		m_jTableBottom.setRowSelectionAllowed(true);
@@ -1022,7 +1001,7 @@ public class PlayerComparePanel extends LazyImagePanel implements ItemListener, 
 			for (int u = 0; u < tmpAnzahl; u++) {
 				tmpPlayer = m_V_setPlayers.elementAt(u);
 				if (("" + tmpPlayer.getId()).compareTo(id) == 0) {
-					l_SpielerName.setText("" + tmpPlayer.getFullName());
+					l_SpielerName.setText(tmpPlayer.getFullName());
 					break;
 				}
 			}