+++ src/main/resources/release_notes.md
@@ -14,7 +14,9 @@
 
 * Calculation of the tsi sub by approximating the tsi formula (#235)
 
-### Team Analyzer
+### Player Analysis
+
+* Fix error in rating compare table (#2153)
 
 ### Rating